move VueAuthenticate plugin loading to before new Vue()
documentation says that it should be in this order
